OC-206: rename test classes as using old&new numbering is confusing
